rf: refactor routes, pages && fix bug
