+++ src/Pages/FAQ/FAQ.js
