+++ src/Pages/ManAccess/ManAccessories.js
@@ -1,4 +1,4 @@
-import MenAccessories from "../Components/ManAccessories/man-accessories.js";
+import MenAccessories from "../../Components/ManAccessories/man-accessories.js.js";
 
 const ManAccessories = () => {
   return (
