+++ src/Pages/PrivacyPolicy/PrivacyPolicy.js
