+++ src/Routes/routes.js
@@ -1,23 +1,23 @@
 import ShopHome from "../Components/Shop/ShopHome/shop-home";
 import Watches from "../Components/Watches/Watches";
-import About from "../Pages/About";
-import Blog from "../Pages/Blog";
-import Contact from "../Pages/Contact";
-import FashionAccessories from "../Pages/FashionAccessories";
-import GetCoupon from "../Pages/GetCoupon";
-import Home from "../Pages/Home";
-import Latest from "../Pages/Latest";
-import ManAccessories from "../Pages/ManAccessories";
-import OffersAndDiscounts from "../Pages/OffersAndDiscounts";
-import PrivacyPolicy from "../Pages/PrivacyPolicy";
-import ReportIssue from "../Pages/ReportIssue";
-import RubberToys from "../Pages/RubberToys";
-import Shop from "../Pages/Shop";
-import TermsAndCondition from "../Pages/TermsAndCondition";
-import WomanCloth from "../Pages/WomanCloth";
-import CartPage from "../Pages/cart-page";
+import About from "../Pages/About/About";
+import Blog from "../Pages/Blog/Blog";
+import Contact from "../Pages/Contact/Contact";
+import FashionAccessories from "../Pages/FashionAccess/FashionAccessories";
+import GetCoupon from "../Pages/Coupons/GetCoupon";
+import Home from "../Pages/Home/Home";
+import ManAccessories from "../Pages/ManAccess/ManAccessories";
+import OffersAndDiscounts from "../Pages/OfferAndDiscount/OffersAndDiscounts";
+import PrivacyPolicy from "../Pages/PrivacyPolicy/PrivacyPolicy";
+import ReportIssue from "../Pages/ReportIssue/ReportIssue";
+import RubberToys from "../Pages/RubberToys/RubberToys";
+import Shop from "../Pages/Shop/Shop";
+import TermsAndCondition from "../Pages/Terms&Condition/TermsAndCondition";
+import WomanCloth from "../Pages/WomanCloth/WomanCloth";
+import CartPage from "../Pages/CartPage/cart-page";
 import ProductDetails from "../Pages/product-details";
-import FAQ from "../Pages/FAQ";
+import FAQ from "../Pages/FAQ/FAQ";
+import Latest from "../Pages/Latest/Latest";
 
 export const routes = [
   {
